lua/nvls/lilypond.lua: adding new quickplayer function

+++ lua/nvls/lilypond.lua
@@ -24,10 +24,10 @@ function M.lilyPlayer()
     local main_last = getLastMod(nvls_main)
 
     if (mp3_last and mp3_last > midi_last) then
-      require('nvls.lilypond').player(lilyAudioFile)
+      require('nvls.lilypond').player(lilyAudioFile, nvls_file_name .. ".mp3")
 
     else
-      print('Converting ' .. nvls_file_name .. '.midi to mp3...') 
+      print('[NVLS] Converting ' .. nvls_file_name .. '.midi to mp3...') 
       local convert = "rm -rf '" .. lilyAudioFile .. "' && " ..
         "fluidsynth -T raw -F - '" .. lilyMidiFile .. 
         "' -s | ffmpeg -f s32le -i - '" .. lilyAudioFile .. "'"
@@ -36,7 +36,7 @@ function M.lilyPlayer()
 
   elseif io.open(fn.glob(main_folder .. '/' .. 
       nvls_short .. '.mp3', "r")) then
-    require('nvls.lilypond').player(lilyAudioFile)
+    require('nvls.lilypond').player(lilyAudioFile, nvls_file_name .. ".mp3")
 
   else
     print("[NVLS] No mp3 file in working directory")
@@ -51,16 +51,14 @@ function M.DefineLilyVars()
 
   if io.open(fn.glob(main_folder .. '/.lilyrc')) then
     dofile(expand(main_folder) .. '/.lilyrc')
-    nvls_main = require('nvls').shellescape(expand(main_folder) .. "/" .. 
-    main_file)
+    nvls_main = require('nvls').shellescape(expand(main_folder) .. "/" .. main_file)
     if not io.open(fn.glob(nvls_main)) then
       nvls_main = require('nvls').shellescape(expand('%:p'))
     end
 
   elseif io.open(fn.glob(expand(main_folder) .. '/' .. 
     main_file)) then
-      nvls_main = require('nvls').shellescape(
-        expand(main_folder) .. "/" .. main_file)
+      nvls_main = require('nvls').shellescape(expand(main_folder) .. "/" .. main_file)
   end
 
   local name,out = nvls_main:gsub("%.(ly)", "")
@@ -74,7 +72,7 @@ function M.DefineLilyVars()
   lilyAudioFile = require('nvls').shellescape(expand(nvls_main_name .. ".mp3"))
 end
 
-function M.player(file)
+function M.player(file,name)
   local lilyPopup = require("nui.popup")
   local plopts = nvls_options.player.options
   
@@ -82,7 +80,7 @@ function M.player(file)
     enter = true,
     focusable = true,
     border = {
-      text = { top = "[" .. nvls_file_name .. ".mp3]" },
+      text = { top = "[" .. name .. "]" },
       style = plopts.border_style,
     },
       position = {
@@ -118,7 +116,7 @@ function M.player(file)
     lilyPlayer:map('n', key, cmd, nrm)
   end
   
-  map(opt.quit,             function() lilyPlayer:unmount() end)
+  map(opt.quit, function() lilyPlayer:unmount() end)
   map(lyopt.switch_buffers, "<cmd>stopinsert<cr><C-w>w")
   map(opt.backward,         "i<Left><cmd>stopinsert<cr>")
   map(opt.forward,          "i<Right><cmd>stopinsert<cr>")
@@ -175,10 +173,8 @@ function M.quickLangInput()
   end, { noremap = true })
 end
 
-function M.getVisualSelection()
+function M.inputString(s_start, s_end)
   local getLines = vim.api.nvim_buf_get_lines
-  local s_start  = vim.fn.getpos("'<")
-  local s_end    = vim.fn.getpos("'>")
   local n_lines  = math.abs(s_end[2] - s_start[2]) + 1
   local lines    = getLines(0, s_start[2] - 1, s_end[2], false)
   lines[1]       = string.sub(lines[1], s_start[3], -1)
@@ -195,26 +191,29 @@ function M.getHyphType()
   if nvls_hyphlang then lang = nvls_hyphlang
   else lang = nvls_options.lilypond.options.hyphenation_language
   end
-  if lang == "en_DEFAULT" then require('nvls.lilypond').hyphenator()
-  else require('nvls.lilypond').pyphen()
+
+  local input = require('nvls.lilypond').inputString(fn.getpos("'<"), fn.getpos("'>"))
+  if lang == "en_DEFAULT" then require('nvls.lilypond').hyphenator(input)
+  else require('nvls.lilypond').pyphen(input)
   end
 end
 
-function M.hyphenator()
-  local input = require('nvls.lilypond').getVisualSelection()
+function M.hyphenator(input)
   require('nvls.hyphs')
   for i, j in pairs(hyphs) do 
     input = input:gsub("%f[%w_]" .. i .. "s?%f[^%w_]", j)
   end
+  fn.execute("set paste")
   fn.execute("normal gvc" .. input)
+  fn.execute("normal g`<")
+  fn.execute("set nopaste")
 end
   
-function M.pyphen()
+function M.pyphen(input)
   if fn.has('python3') == 0 then
     print('[NVLS] python3 is not available')
     do return end
   end
-  local input = require('nvls.lilypond').getVisualSelection()
   input = input:gsub("[\n\r]", " ")
   fn.execute('py3 import pyphen')
   fn.execute('py3 import vim')
@@ -229,22 +228,96 @@ function M.pyphen()
   fn.execute("'<,'>py3do return py_vim_string_replace(line)")
 end
 
--- WORK IN PROGRES
--- MAKE FUNCTION TO PLAY VISUAL SELECTION
+function M.quickplayerInputType()
+  local sel = require('nvls.lilypond').inputString(fn.getpos("'<"), fn.getpos("'>"))
+  local from_top = require('nvls.lilypond').inputString({0, 1, 1, 0}, fn.getpos("'<"))
 
-function M.tempLy()
+  if string.find(sel, "%pfixed%s+%a*%p*%s*%{") or string.find(sel, "%prelative%s+%a*%p*%s*%{") or string.find(sel, "%pchords.*%{") then
+    return ''
 
-  local sel = require('nvls.lilypond').getVisualSelection()
+  elseif string.find(from_top, "%prelative%s+%a*%p*%s*%{") then
+    from_relative = from_top:gsub(".*%prelative%s+%a*%p*%s*%{", "")
 
+    if string.find(from_relative, "%pfixed%s+%a*%p*%s*%{") then
+      from_fixed = from_relative:gsub(".*pfixed%s+%a*%p*%s*%{", "")
+      if string.find(from_fixed, "%pchords%s+%{") then
+        return "\\chords"
+      else
+        local input_type = "\\fixed"
+        local ref_pitch = string.match(from_relative, ".*%pfixed(%s+%a%p*)")
+        return input_type .. ref_pitch
+      end
+    elseif string.find(from_relative, "%pchords%s+%{") then
+        return "\\chords"
+    else
+      local input_type = "\\relative"
+      local ref_pitch = string.match(from_top, ".*%prelative(%s+%a%p*)")
+      return input_type .. ref_pitch
+    end
+
+  elseif string.find(from_top, "%pfixed%s+%a*%p*%s*%{") then
+    from_fixed = from_top:gsub(".*%pfixed%s+%a*%p*%s*%{", "")
+
+    if string.find(from_fixed, "%prelative%s+%a*%p*%s*%{") then
+      from_relative = from_fixed:gsub(".*prelative%s+%a*%p*%s*%{", "")
+      if string.find(from_relative, "%pchords%s+%{") then
+        return "\\chords"
+      else
+        local input_type = "\\relative"
+        local ref_pitch = string.match(from_fixed, ".*%prelative(%s+%a%p*)")
+        return input_type .. ref_pitch
+      end
+    elseif string.find(from_fixed, "%pchords%s+%{") then
+        return "\\chords"
+    else
+      local input_type = "\\fixed"
+      local ref_pitch = string.match(from_top, ".*%pfixed(%s+%a%p*)")
+      return input_type .. ref_pitch
+    end
+
+  elseif string.find(from_top, "%pchords%s+%{") then
+    from_chords = from_top:gsub(".*%pchords%s+%{", "")
+
+    if string.find(from_chords, "%prelative%s+%a*%p*%s*%{") then
+      from_relative = from_chords:gsub(".*prelative%s+%a*%p*%s*%{", "")
+      if string.find(from_relative, "%pfixed%s+%a*%p*%s*%{") then
+        local input_type = "\\fixed"
+        local ref_pitch = string.match(from_chords, ".*%pfixed(%s+%a%p*)")
+        return input_type .. ref_pitch
+      else
+        local input_type = "\\relative"
+        local ref_pitch = string.match(from_chords, ".*%prelative(%s+%a%p*)")
+        return input_type .. ref_pitch
+      end
+    elseif string.find(from_chords, "%pfixed%s+%a*%p*%s*%{") then
+      from_fixed = from_chords:gsub(".*pfixed%s+%a*%p*%s*%{", "")
+      if string.find(from_fixed, "%prelative%s+%a*%p*%s*%{") then
+        local input_type = "\\relative"
+        local ref_pitch = string.match(from_chords, ".*%prelative(%s+%a%p*)")
+        return input_type .. ref_pitch
+      else
+        local input_type = "\\fixed"
+        local ref_pitch = string.match(from_chords, ".*%pfixed(%s+%a%p*)")
+        return input_type .. ref_pitch
+      end
+    else
+        return "\\chords"
+    end
+  else
+    print("[NVLS] Can't find input type") do return end
+  end
+end
+
+function M.quickplayerCheckErr(string)
   local function countChar(str,char)
     local _,n = str:gsub(char,"")
     return n
   end
 
-  local op_curl_br  = countChar(sel,[[{]])
-  local cl_curl_br  = countChar(sel,[[}]])
-  local op_angle_br = countChar(sel,[[<]])
-  local cl_angle_br = countChar(sel,[[>]])
+  local op_curl_br  = countChar(string,[[{]])
+  local cl_curl_br  = countChar(string,[[}]])
+  local op_angle_br = countChar(string,[[<]])
+  local cl_angle_br = countChar(string,[[>]])
 
   if op_curl_br ~= cl_curl_br then
     print('[NVLS] Curly brackets not matching in visual selection')
@@ -252,9 +325,25 @@ function M.tempLy()
   elseif op_angle_br ~= cl_angle_br then
     print('[NVLS] Angle brackets not matching in visual selection')
     do return end
+  elseif string.find(string, "%pscore%s") then
+    print("[NVLS] Can't compile with \\score in visual selection")
+    do return end
+  else
+    return "OK"
+  end
+end
+
+function M.quickplayer()
+  local sel = require('nvls.lilypond').inputString(fn.getpos("'<"), fn.getpos("'>"))
+
+  if require('nvls.lilypond').quickplayerCheckErr(sel) ~= "OK" then
+    do return end
   end
 
-  local code = "\\score { \\relative c' { " .. sel .. " } \\midi {} }"
+  local input_type = require('nvls.lilypond').quickplayerInputType()
+  local code = "\\score { " .. input_type .. " { " .. sel .. " } \\midi {} }"
+
+  print("[NVLS] Converting to mp3...")
   tmpOutDir = "/tmp/nvls"
   os.execute('rm -rf ' .. tmpOutDir)
   os.execute('mkdir -p ' .. tmpOutDir)
@@ -262,6 +351,7 @@ function M.tempLy()
   tmpfile:write(code)
   tmpfile:close()
   os.execute('lilypond --loglevel=NONE -o ' .. tmpOutDir .. ' ' .. tmpOutDir .. '/tmp.ly')
+
   local convert = 'fluidsynth -T raw -F - ' .. tmpOutDir .. '/tmp.midi' ..
       ' -s | ffmpeg -f s32le -i - ' .. tmpOutDir .. '/tmp.mp3'
   require('nvls').make(convert," ","tmpplayer")
